+++ src/custom/cards/sets/SetDefinition_Demo1.gd
@@ -13,6 +13,7 @@ const CARDS := {
 		"Cost": 0,
 		"Power": 0,
 		"_max_allowed": 1,
+		"_illustration": "Nobody"
 	},
 	"Test Card 2": {
 		"Type": "Red",
@@ -25,6 +26,7 @@ const CARDS := {
 		"Cost": 2,
 		"Power": 5,
 		"_max_allowed": 10,
+		"_illustration": "Nobody"
 	},
 	"Spawn Card": {
 		"Type": "Token",